Backend/brain-service/requirements.txt
# Core dependencies
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
python-dotenv==1.0.1
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic-settings==2.1.0
httpx==0.26.0

# ML and AI
langchain==0.0.335
openai==1.12.0
transformers==4.34.0
torch==2.1.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1

# Financial analysis
yfinance==0.2.31
ta-lib==0.4.28
pandas-ta==0.3.14b0
ccxt==4.1.13

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1 